requirements.txt
openai
pytrends
tweepy
python-dotenv
